backend/requirements.txt
aniso8601==9.0.1
click==8.0.1
Faker==8.12.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.2
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
text-unidecode==1.3
Werkzeug==2.0.1
